# perf(functions): Optimizations for join()