--- a/packages/functions/src/compact-primitive.ts
+++ b/packages/functions/src/compact-primitive.ts
@@ -1,5 +1,5 @@
 import { Accessor, Document, Primitive, TypedArray, TypedArrayConstructor } from '@gltf-transform/core';
-import { createIndicesEmpty, deepListAttributes, shallowCloneAccessor } from './utils.js';
+import { createIndices, createIndicesEmpty, deepListAttributes, shallowCloneAccessor } from './utils.js';
 import { VertexCountMethod, getPrimitiveVertexCount } from './get-vertex-count.js';
 import { EMPTY_U32 } from './hash-table.js';
 
@@ -7,14 +7,8 @@ import { EMPTY_U32 } from './hash-table.js';
 export function compactPrimitive(prim: Primitive, remap?: TypedArray, dstVertexCount?: number): Primitive {
 	const document = Document.fromGraph(prim.getGraph())!;
 
-	// Compute remap and dstVertexCount if not provided.
-
 	if (!remap || !dstVertexCount) {
-		const compactPlan = createCompactPlan(prim);
-		if (!compactPlan) {
-			return prim; // Unindexed; skip.
-		}
-		[remap, dstVertexCount] = compactPlan;
+		[remap, dstVertexCount] = createCompactPlan(prim);
 	}
 
 	// Remap indices.
@@ -102,16 +96,19 @@ export function compactAttribute(
 /**
  * Creates a 'remap' and 'dstVertexCount' plan for indexed primitives,
  * such that they can be rewritten with {@link compactPrimitive} removing
- * any non-rendered vertices. For unindexed primitives, returns null.
+ * any non-rendered vertices.
  * @hidden
  * @internal
  */
-function createCompactPlan(prim: Primitive): [Uint32Array, number] | null {
+function createCompactPlan(prim: Primitive): [Uint32Array, number] {
+	const srcVertexCount = getPrimitiveVertexCount(prim, VertexCountMethod.UPLOAD);
+
 	const indices = prim.getIndices();
 	const indicesArray = indices ? indices.getArray() : null;
-	if (!indices || !indicesArray) return null;
+	if (!indices || !indicesArray) {
+		return [createIndices(srcVertexCount, 1_000_000) as Uint32Array, srcVertexCount];
+	}
 
-	const srcVertexCount = getPrimitiveVertexCount(prim, VertexCountMethod.UPLOAD);
 	const remap = new Uint32Array(srcVertexCount).fill(EMPTY_U32);
 
 	let dstVertexCount = 0;