--- a/packages/functions/src/index.ts
+++ b/packages/functions/src/index.ts
@@ -1,6 +1,7 @@
 export * from './center.js';
 export * from './clear-node-parent.js';
 export * from './clear-node-transform.js';
+export * from './compact-primitive.js';
 export * from './convert-primitive-mode.js';
 export * from './dedup.js';
 export { dequantize, dequantizePrimitive, DequantizeOptions } from './dequantize.js';