--- a/packages/functions/src/join.ts
+++ b/packages/functions/src/join.ts
@@ -15,6 +15,8 @@ import { prune } from './prune.js';
 import { transformPrimitive } from './transform-primitive.js';
 import { assignDefaults, createPrimGroupKey, createTransform, formatLong, isUsed } from './utils.js';
 import { dequantizeAttribute } from './dequantize.js';
+import { getMeshVertexCount, getPrimitiveVertexCount, VertexCountMethod } from './get-vertex-count.js';
+import { compactPrimitive } from './compact-primitive.js';
 
 const NAME = 'join';
 
@@ -156,6 +158,12 @@ function _joinLevel(document: Document, parent: Node | Scene, options: Required<
 			const material = prim.getMaterial();
 			if (material && material.getExtension('KHR_materials_volume')) continue;
 
+			const renderVertexCount = getPrimitiveVertexCount(prim, VertexCountMethod.RENDER);
+			const uploadVertexCount = getPrimitiveVertexCount(prim, VertexCountMethod.UPLOAD);
+			if (renderVertexCount < uploadVertexCount / 2) {
+				compactPrimitive(prim);
+			}
+
 			dequantizeTransformableAttributes(prim);
 
 			let key = createPrimGroupKey(prim);
@@ -216,7 +224,8 @@ function _joinLevel(document: Document, parent: Node | Scene, options: Required<
 			primMesh.removePrimitive(prim);
 
 			// Primitives may be reused directly, or their attributes may be
-			// used in another Primitive with a different Material.
+			// used in another Primitive with a different Material. Because
+			// compactPrimitive ran already, >=50% of vertices are used.
 			if (isUsed(prim) || hasSharedAttributes(prim)) {
 				prim = prims[i] = _deepClonePrimitive(prims[i]);
 			}
@@ -224,11 +233,11 @@ function _joinLevel(document: Document, parent: Node | Scene, options: Required<
 			// Transform Primitive into new local coordinate space.
 			if (primNode !== dstNode) {
 				multiply(_matrix, invert(_matrix, dstMatrix), primNode.getMatrix());
-				transformPrimitive(prim, _matrix);
+				transformPrimitive(prim, _matrix); // TODO(perf)
 			}
 		}
 
-		const dstPrim = joinPrimitives(prims);
+		const dstPrim = joinPrimitives(prims); // TODO(perf)
 		const dstVertexCount = dstPrim.listAttributes()[0].getCount();
 		dstMesh.addPrimitive(dstPrim);
 
@@ -240,6 +249,7 @@ function _joinLevel(document: Document, parent: Node | Scene, options: Required<
 }
 
 function _deepClonePrimitive(src: Primitive): Primitive {
+	// compactPrimitive already applied, we know >=50% of vertices are used.
 	const dst = src.clone();
 	for (const semantic of dst.listSemantics()) {
 		dst.setAttribute(semantic, dst.getAttribute(semantic)!.clone());