--- a/packages/functions/src/join.ts
+++ b/packages/functions/src/join.ts
@@ -15,7 +15,6 @@ import { prune } from './prune.js';
 import { transformPrimitive } from './transform-primitive.js';
 import { assignDefaults, createPrimGroupKey, createTransform, formatLong, isUsed } from './utils.js';
 import { dequantizeAttribute } from './dequantize.js';
-import { getMeshVertexCount, getPrimitiveVertexCount, VertexCountMethod } from './get-vertex-count.js';
 import { compactPrimitive } from './compact-primitive.js';
 
 const NAME = 'join';
@@ -218,10 +217,11 @@ function _joinLevel(document: Document, parent: Node | Scene, options: Required<
 			let prim = prims[i];
 			primMesh.removePrimitive(prim);
 
-			// Primitives may be reused directly, or their attributes may be
-			// used in another Primitive with a different Material. Because
-			// compactPrimitive ran already, 100% of vertices are used.
-			if (isUsed(prim) || hasSharedAttributes(prim)) {
+			// If Primitive is still in use after being removed from the
+			// current mesh, above, make a deep copy. Because compactPrimitive()
+			// was applied earlier in join(), we know the full vertex streams are
+			// used, and no accessors are shared.
+			if (isUsed(prim)) {
 				prim = prims[i] = _deepClonePrimitive(prims[i]);
 			}
 
@@ -244,7 +244,7 @@ function _joinLevel(document: Document, parent: Node | Scene, options: Required<
 }
 
 function _deepClonePrimitive(src: Primitive): Primitive {
-	// compactPrimitive already applied, we know >=50% of vertices are used.
+	// compactPrimitive already applied; no vertices are unused.
 	const dst = src.clone();
 	for (const semantic of dst.listSemantics()) {
 		dst.setAttribute(semantic, dst.getAttribute(semantic)!.clone());
@@ -252,17 +252,6 @@ function _deepClonePrimitive(src: Primitive): Primitive {
 	const indices = dst.getIndices();
 	if (indices) dst.setIndices(indices.clone());
 	return dst;
-}
-
-function hasSharedAttributes(prim: Primitive): boolean {
-	for (const attribute of prim.listAttributes()) {
-		for (const parent of attribute.listParents()) {
-			if (parent !== prim && parent.propertyType !== ROOT) {
-				return true;
-			}
-		}
-	}
-	return false;
 }
 
 /**