--- a/packages/functions/src/transform-mesh.ts
+++ b/packages/functions/src/transform-mesh.ts
@@ -1,6 +1,6 @@
-import { mat4, Accessor, Primitive, Mesh, PropertyType, PrimitiveTarget } from '@gltf-transform/core';
+import { mat4, Mesh } from '@gltf-transform/core';
 import { transformPrimitive } from './transform-primitive.js';
-import { deepListAttributes } from './utils.js';
+import { compactPrimitive } from './compact-primitive.js';
 
 /**
  * Applies a transform matrix to every {@link Primitive} in the given {@link Mesh}.
@@ -28,61 +28,24 @@ import { deepListAttributes } from './utils.js';
  * @param matrix
  * @param overwrite Whether to overwrite vertex streams in place. If false,
  * 		streams shared with other meshes will be detached.
- * @param skipIndices Vertices, specified by index, to be _excluded_ from the transformation.
  */
-export function transformMesh(mesh: Mesh, matrix: mat4, overwrite = false, skipIndices?: Set<number>): void {
-	// (1) Detach shared prims.
-	for (const srcPrim of mesh.listPrimitives()) {
-		const isShared = srcPrim.listParents().some((p) => p.propertyType === PropertyType.MESH && p !== mesh);
-		if (isShared) {
-			const dstPrim = srcPrim.clone();
-			mesh.swap(srcPrim, dstPrim);
-
-			for (const srcTarget of dstPrim.listTargets()) {
-				const dstTarget = srcTarget.clone();
-				dstPrim.swap(srcTarget, dstTarget);
-			}
-		}
-	}
-
-	// (2) Detach shared vertex streams.
-	if (!overwrite) {
-		const parents = new Set<Primitive | PrimitiveTarget>([
-			...mesh.listPrimitives(),
-			...mesh.listPrimitives().flatMap((prim) => prim.listTargets()),
-		]);
-		const attributes = new Map<Accessor, Accessor>();
-		for (const prim of mesh.listPrimitives()) {
-			for (const srcAttribute of deepListAttributes(prim)) {
-				const isShared = srcAttribute
-					.listParents()
-					.some((a) => (a instanceof Primitive || a instanceof PrimitiveTarget) && !parents.has(a));
-				if (isShared && !attributes.has(srcAttribute)) {
-					attributes.set(srcAttribute, srcAttribute.clone());
-				}
-			}
-		}
-		for (const parent of parents) {
-			for (const [srcAttribute, dstAttribute] of attributes) {
-				parent.swap(srcAttribute, dstAttribute);
-			}
-		}
-	}
-
-	// (3) Apply transform.
-	const attributeSkipIndices = new Map<Accessor, Set<number>>();
+export function transformMesh(mesh: Mesh, matrix: mat4): void {
 	for (const prim of mesh.listPrimitives()) {
-		const position = prim.getAttribute('POSITION')!;
-
-		let primSkipIndices;
-		if (skipIndices) {
-			primSkipIndices = skipIndices;
-		} else if (attributeSkipIndices.has(position)) {
-			primSkipIndices = attributeSkipIndices.get(position)!;
-		} else {
-			attributeSkipIndices.set(position, (primSkipIndices = new Set<number>()));
-		}
-
-		transformPrimitive(prim, matrix, primSkipIndices);
+		compactPrimitive(prim);
+		transformPrimitive(prim, matrix);
 	}
+
+	// const attributeSkipIndices = new Map<Accessor, Uint32Array>();
+	// for (const prim of mesh.listPrimitives()) {
+	// 	const position = prim.getAttribute('POSITION')!;
+
+	// 	let primSkipIndices: Uint32Array;
+	// 	if (attributeSkipIndices.has(position)) {
+	// 		primSkipIndices = attributeSkipIndices.get(position)!;
+	// 	} else {
+	// 		attributeSkipIndices.set(position, (primSkipIndices = new Uint32Array(position.getCount())));
+	// 	}
+
+	// 	transformPrimitive(prim, matrix, primSkipIndices);
+	// }
 }