Changes:

- BREAKING CHANGE: Removes 'overwrite' and 'skipIndices' options from `transformMesh()` and `transformPrimitive()` functions
- feat(functions): Exposes compactPrimitive as public API
- perf(functions): Improves `join()` and `joinPrimitives()` performance
- perf(functions): Improves `transformMesh()` `transformPrimitive()` performance

Related:

- Fixes https://github.com/donmccurdy/glTF-Transform/issues/1315
- https://github.com/donmccurdy/glTF-Transform/issues/1133

Benchmark:

```plaintext
join
v3.10.1             ▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀     127.446ms
v4.0.0-alpha.12     ▀▀▀▀▀▀▀▀▀▀▀                                  33.6132ms
v4.0.0-alpha.13     ▀▀▀▀▀▀▀▀▀▀▀                                  36.136ms
v4.0.0-alpha.14     ▀▀▀▀▀▀▀▀▀▀▀▀                                 39.452ms
v4.0.0-alpha.15     ▀▀▀▀▀▀▀▀▀▀▀                                  33.7321ms
dev                 ▀▀▀▀▀▀                                       18.5413ms
```